avoid rerendes with use travel path hook
