+++ client/src/ui/components/worldmap/hexagon/WorldHexagon.tsx
@@ -4,18 +4,11 @@ import { Flags } from "@/ui/components/worldmap/Flags.jsx";
 import { Subscription } from "rxjs";
 import { create } from "zustand";
 import { useDojo } from "../../../../hooks/context/DojoContext";
-import useUIStore from "../../../../hooks/store/useUIStore.js";
 import { ShardsMines } from "../../models/buildings/worldmap/ShardsMines.js";
 import { Structures } from "../../models/buildings/worldmap/Structures.js";
 import { Armies } from "../armies/Armies.js";
 import { BiomesGrid, HexagonGrid } from "./HexLayers.js";
-
-import { useStamina } from "@/hooks/helpers/useStamina";
-import { HighlightPositions } from "@/types";
-import { ACCESSIBLE_POSITIONS_COLOUR, COLS, FELT_CENTER, ROWS } from "@/ui/config.js";
-import { getUIPositionFromColRow } from "@/ui/utils/utils";
-import { EternumGlobalConfig } from "@bibliothecadao/eternum";
-import { findAccessiblePositionsAndPaths } from "./utils.js";
+import { COLS, FELT_CENTER, ROWS } from "@/ui/config.js";
 import { Battles } from "../../models/buildings/worldmap/Battles";
 
 interface ExploredHexesState {
@@ -102,44 +95,7 @@ export const WorldMap = () => {
     };
   }, [setExploredHexes, exploreMapEvents]);
 
-  const selectedEntity = useUIStore((state) => state.selectedEntity);
-  const setHighlightPositions = useUIStore((state) => state.setHighlightPositions);
-  const setTravelPaths = useUIStore((state) => state.setTravelPaths);
-
-  const { useStaminaByEntityId } = useStamina();
-  const stamina = useStaminaByEntityId({ travelingEntityId: selectedEntity?.id || 0n });
-
-  useEffect(() => {
-    if (!selectedEntity || !stamina) return;
-
-    const maxTravelPossible = Math.floor((stamina.amount || 0) / EternumGlobalConfig.stamina.travelCost);
-    const canExplore = (stamina.amount || 0) >= EternumGlobalConfig.stamina.exploreCost;
-
-    const pathMap = findAccessiblePositionsAndPaths(
-      selectedEntity.position,
-      exploredHexes,
-      maxTravelPossible,
-      canExplore,
-    );
-
-    const path = Array.from(pathMap.entries()).map(([key, path]) => {
-      const [x, y] = key.split(",").map(Number);
-      return { x, y, path };
-    });
-
-    if (path.length <= 1) return;
-
-    const uiPath: HighlightPositions = {
-      pos: path.map(({ x, y }) => {
-        const pos = getUIPositionFromColRow(x, y);
-        return [pos.x, -pos.y];
-      }),
-      color: ACCESSIBLE_POSITIONS_COLOUR,
-    };
-
-    setHighlightPositions(uiPath);
-    setTravelPaths(pathMap);
-  }, [selectedEntity, stamina, exploredHexes]);
+  console.log("hello");
 
   return (
     <>
